requirements.txt
PyPDF2
numpy
langchain 
langchain-community
openai
sentence-transformers
transformers
faiss-cpu
google-generativeai
python-dotenv
streamlit



